+++ app/controllers/admin/claims_controller.rb
@@ -2,17 +2,15 @@ class Admin::ClaimsController < Admin::BaseAdminController
   before_action :ensure_service_operator
 
   def index
-    claims = Claim.includes(eligibility: [:claim_school, :current_school]).submitted.order(:submitted_at)
-    csv = TslrClaimsCsv.new(claims)
+    @claims = Claim.includes(eligibility: [:claim_school, :current_school]).awaiting_approval.order(:submitted_at)
 
     respond_to do |format|
-      format.csv { send_file csv.file, type: "text/csv", filename: "claims.csv" }
+      format.csv { send_file TslrClaimsCsv.new(@claims).file, type: "text/csv", filename: "claims.csv" }
+      format.html
     end
   end
 
-  private
-
-  def ensure_service_operator
-    render "admin/auth/failure", status: :unauthorized unless service_operator_signed_in?
+  def show
+    @claim = Claim.find(params[:id])
   end
 end