+++ db/migrate/20190905140712_add_approved_at_to_claim.rb
@@ -1,0 +1,5 @@
+class AddApprovedAtToClaim < ActiveRecord::Migration[5.2]
+  def change
+    add_column :claims, :approved_at, :datetime
+  end
+end