+++ db/migrate/20190906082123_add_approved_by_to_claim.rb
@@ -1,0 +1,5 @@
+class AddApprovedByToClaim < ActiveRecord::Migration[5.2]
+  def change
+    add_column :claims, :approved_by, :string
+  end
+end